backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.27.1
pydantic>=2.11.0
pillow>=10.4.0

